Do not strip minor version off _REVISION
Introduced in commit 613fe532, _REVISION was redefined to be just the
major version iff PKG_VERSION was not overridden (default) and the
branch was -CURRENT, -STABLE, or -PRERELEASE.  This introduced side
effects in other parts of the code which expected _REVISION to be
major.minor.

This commit introduces MAJOR_REVISION to alleviate this problem:
_REVISION is once again major.minor, and existing uses of _REVISION now
see major.minor again, except PKG_VERSION (14.snap20220311121531) and
PKG_ABI (FreeBSD:14:amd64) need only the major version and now use
MAJOR_VERSION instead.

This commit also fixes the bug where _REVISION was left in the
major.minor format and broke PKG_ABI iff PKG_VERSION was overridden:
PKG_VERSION was using the ${_REVISION:S/\..*$//} make(1) substitution
to drop minor version if any, which did not work because :S does not
accept regex (:C does).